Chore : AuthController에서 UserController 로 닉네임 관련 API 이동 (#5)
